fix(FOROME-904): sort folders structure
