+++ src/pages/ws/ui/control-panel/control-panel.tsx
@@ -4,20 +4,18 @@ import { t } from '@i18n'
 import zoneStore from '@store/filterZone'
 import { ControlPanelDivider } from './control-panel-divider'
 import { EditFilter } from './control-panel-edit-filter'
-import { FilterItemGenes } from './control-panel-filter-genes'
-import { FilterItemGenesList } from './control-panel-filter-genes-list'
-import { FilterItemSamples } from './control-panel-filter-samples'
-import { FilterItemTags } from './control-panel-filter-tags'
 import { ControlPanelPreset } from './control-panel-preset'
-import { Results } from './control-panel-settings'
+import { CustomizeTable } from './customize-table-modal/customize-table'
 import { DatasetCreationButton } from './dataset-creation-button'
-import { GenesListModal } from './genes-list-modal'
-import { GenesModal } from './genes-modal'
-import { ZoneItem } from './zone-item'
+import { GenesListModal } from './zone-modals/genes-list-modal'
+import { GenesModal } from './zone-modals/genes-modal'
+import { SamplesModal } from './zone-modals/samples-modal'
+import { TagsModal } from './zone-modals/tags-modal'
+import { ZoneItem } from './zone-modals/zone-item'
 
 export const ControlPanel = (): ReactElement => (
   <div className="w-auto flex px-4 bg-blue-dark">
-    <div className="rounded flex px-4 pt-2 pb-4">
+    <div className="rounded flex pt-2 pb-4">
       <ControlPanelPreset />
 
       <ControlPanelDivider />
@@ -26,7 +24,7 @@ export const ControlPanel = (): ReactElement => (
 
       <ControlPanelDivider />
 
-      <Results />
+      <CustomizeTable />
 
       <div className="flex ml-5 bg-blue-darkHover rounded-sm px-3">
         <ZoneItem
@@ -41,15 +39,17 @@ export const ControlPanel = (): ReactElement => (
           data={zoneStore.selectedGenesList}
         />
 
-        {/* <FilterItemGenesList title={t('ds.geneList')} /> */}
+        <ZoneItem
+          title={t('ds.sample')}
+          modalElement={SamplesModal}
+          data={zoneStore.selectedSamples}
+        />
 
-        {/* <ControlPanelDivider className="bg-blue-secondary" />
-
-        <FilterItemSamples title={t('ds.sample')} />
-
-        <ControlPanelDivider className="bg-blue-secondary" />
-
-        <FilterItemTags title={t('ds.tags')} /> */}
+        <ZoneItem
+          title={t('ds.tags')}
+          modalElement={TagsModal}
+          data={zoneStore.selectedTags}
+        />
       </div>
 
       <DatasetCreationButton />
