+++ src/pages/ws/ui/control-panel/customize-table-modal/columns-list/columns-list.tsx
