+++ src/pages/ws/ui/control-panel/customize-table-modal/columns-list/components/column-name-item.tsx
