+++ src/pages/ws/ui/control-panel/customize-table-modal/customize-table.tsx
@@ -1,0 +1,14 @@
+import { ReactElement } from 'react'
+
+import { PopperButton } from '@components/popper-button'
+import { CustomizeTableButton } from './customize-table-button'
+import { CustomizeTableModal } from './customize-table-modal'
+
+export const CustomizeTable = (): ReactElement => (
+  <div>
+    <PopperButton
+      ButtonElement={CustomizeTableButton}
+      ModalElement={CustomizeTableModal}
+    />
+  </div>
+)
